services/vehicles/requirements.txt
Flask==1.0.2
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.4
Flask-Testing==0.6.2
requests
